+++ apps/nextjs/src/app/(main)/@navbar/default.tsx
@@ -1,7 +1,13 @@
+import { Suspense } from "react";
+import { auth } from "@clerk/nextjs/server";
+
+import { cn } from "@battle-stadium/ui";
+
 import NavbarContainer from "~/app/(main)/@navbar/_components/navbar-container";
-import NavbarLinks from "~/app/(main)/@navbar/_components/navbar-links";
 import RightMenu from "~/app/(main)/@navbar/_components/navbar-right-menu";
 import BattleStadium from "~/components/battle-stadium";
+import { NavbarItemsConfigs } from "~/lib/config/site";
+import NavbarLink from "./_components/navbar-link";
 
 export default function Navbar() {
   return (
@@ -13,5 +19,36 @@ export default function Navbar() {
       <NavbarLinks />
       <RightMenu />
     </NavbarContainer>
+  );
+}
+
+function NavbarLinks() {
+  return (
+    <div className="hidden items-center justify-center md:flex">
+      {NavbarItemsConfigs.map(({ label, value }) => (
+        <NavbarLink
+          key={value}
+          value={value}
+          href={`/${value}`}
+          label={label}
+        />
+      ))}
+      <Suspense fallback={null}>
+        <DashboardNavLink />
+      </Suspense>
+    </div>
+  );
+}
+
+async function DashboardNavLink() {
+  const { sessionId } = await auth();
+  return (
+    <NavbarLink
+      value="dashboard"
+      key="dashboard"
+      href="/dashboard"
+      className={cn("items-center", sessionId ? "md:flex" : "hidden")}
+      label="Dashboard"
+    />
   );
 }
